Add Goals for Execs
New function to allow the WordPress admin to set a Goal integer.

This can be done by the 'Show Goals' button which has been added to the top right of the my-events pages

The Goal number will be displayed for Execs next to the 'My Submitted Events' title on the my-events page.

The Goal can also be displayed anywhere else through the shortcode:
[show_goals]